[refactor] : change constants class location
