+++ src/main/java/racingcar/domain/Car.java
@@ -1,6 +1,5 @@
 package racingcar.domain;
 
-import racingcar.utils.Constants;
 import racingcar.utils.RandomNumber;
 
 import java.util.stream.Collectors;
